main,etc to /src, q_obj.empty erased, plan working
