+++ render/get_closest.c
@@ -6,7 +6,7 @@
 /*   By: mpons <mpons@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/08 15:46:21 by slott             #+#    #+#             */
-/*   Updated: 2022/09/26 20:34:11 by mpons            ###   ########.fr       */
+/*   Updated: 2022/09/30 20:40:37 by mpons            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,22 +18,19 @@ void	get_closest_sp(t_set *set, t_ray r)
 	float	t;
 
 	i = 0;
-	if (set->q_obj.sp > 0)
+	while (i < set->q_obj.sp)
 	{
-		while (set->sp_list[i].empty == 0)
+		t = hit_sp(set->sp_list[i], r);
+		if (t < set->obj.dist && t > 0)
 		{
-			t = hit_sp(set->sp_list[i], r);
-			if (t < set->obj.dist && t > 0)
-			{
-				set->obj.dist = t;
-				set->obj.idx = i;
-				set->obj.type = SPHERE;
-				set->obj.col = set->sp_list[i].rgb;
-				set->normal = unit_vector(moins(2, point_at(r, t), \
-							set->sp_list[i].c));
-			}
-			i++;
+			set->obj.dist = t;
+			set->obj.idx = i;
+			set->obj.type = SPHERE;
+			set->obj.col = set->sp_list[i].rgb;
+			set->normal = unit_vector(moins(2, point_at(r, t), \
+						set->sp_list[i].c));
 		}
+		i++;
 	}
 }
 
@@ -42,18 +39,15 @@ void	get_closest_cyl(t_set *set, t_ray r)
 	int		i;
 
 	i = 0;
-	if (set->q_obj.cy > 0)
+	while (i < set->q_obj.cy)
 	{
-		while (set->cyl_list[i].empty == 0)
+		if (check_cyl(set, set->cyl_list[i], r))
 		{
-			if (check_cyl(set, set->cyl_list[i], r))
-			{
-				set->obj.idx = i;
-				set->obj.type = CYLINDRE;
-				set->obj.col = set->cyl_list[i].rgb;
-			}
-			i++;
+			set->obj.idx = i;
+			set->obj.type = CYLINDRE;
+			set->obj.col = set->cyl_list[i].rgb;
 		}
+		i++;
 	}
 }
 
@@ -63,29 +57,21 @@ void	get_closest_pl(t_set *set, t_ray r)//, int flag)
 	float	t;
 
 	i = 0;
-	// while (i < q_obj.pl)
-	if (set->q_obj.pl > 0)
+	while (i < set->q_obj.pl)
 	{
-		while (set->plan_list[i].empty == 0)
+		t = hit_plan(set->plan_list[i], r);
+		if (t < set->obj.dist && t > 0)
 		{
-			t = hit_plan(set->plan_list[i], r);
-			if (t < set->obj.dist && t >= 0)
-			{
-				set->obj.dist = t;
-				set->obj.idx = i;
-				set->obj.type = PLAN;
-				set->obj.col = set->plan_list[i].rgb;
-				if (dot(set->plan_list[i].dir, r.dir) < 0)
-				{
-					// print_vec(set->plan_list[i].dir);
-			    	set->normal = invert_vector(set->plan_list[i].dir);
-					// print_vec(set->normal);
-				}
-				else
-					set->normal = set->plan_list[i].dir;
-			}
-			i++;
+			set->obj.dist = t;
+			set->obj.idx = i;
+			set->obj.type = PLAN;
+			set->obj.col = set->plan_list[i].rgb;
+			if (dot(set->plan_list[i].dir, r.dir) > 0)
+				set->normal = invert_vector(set->plan_list[i].dir);
+			else
+				set->normal = set->plan_list[i].dir;
 		}
+		i++;
 	}
 }
 
@@ -102,9 +88,6 @@ t_obj	init_obj(float t_max)
 
 int	get_closest(t_set *set, t_ray r, float t_max)
 {
-	int	i;
-
-	i = 0;
 	set->obj = init_obj(t_max);
 	get_closest_sp(set, r);
 	get_closest_pl(set, r);
