+++ render/light.c
@@ -6,7 +6,7 @@
 /*   By: mpons <mpons@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/09 13:28:54 by slott             #+#    #+#             */
-/*   Updated: 2022/09/23 23:00:24 by mpons            ###   ########.fr       */
+/*   Updated: 2022/09/30 20:11:08 by mpons            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
