+++ scenes/aNO.rt
@@ -34,13 +34,18 @@ sp	0,10,-10	18				0,255,255
 
 #	Plan
 #	point		normal			color
-pl	8,0,0		-1,0,0			0,0,255
+# Jaune gauche pas
 pl	-7,0,0		1,0,0			255,255,0
 
-# pl	8,0,0		1,0,0			0,0,255
-# pl	-7,0,0		-1,0,0			255,255,0
+#  Rouge fond ver la camera
+# pl 	0,0,-30		0,0,-1			255,0,0 
+#  Rouge fond PAS ver la camera
+pl 	0,0,-30		0,0,1			255,0,0 
 
-pl 	0,0,-30		0,0,-1			40,50,225 
+# Blue droite pas
+pl	8,0,0		-1,0,0			0,0,255
+
+
 # pl 	0,0,-30		0,0,-1			0,0,0 
 # pl 	0,-1,0		0,1,0			0,255,0 
 # pl	6,0,0			-1,0,0		255,32,43
