+++ scenes/aSI.rt
@@ -34,13 +34,23 @@ sp	0,10,-10	18				0,255,255
 
 #	Plan
 #	point		normal			color
-pl	8,0,0		-1,0,0			0,0,255
-pl	-7,0,0		1,0,0			255,255,0
+# pl	8,0,0		-1,0,0			0,0,255
+# pl	-7,0,0		1,0,0			255,255,0
 
-# pl	8,0,0		1,0,0			0,0,255
-# pl	-7,0,0		-1,0,0			255,255,0
+# Jaune gauche vers la camera
+pl	-7,0,0		-1,0,0			255,255,0
 
-pl 	0,0,-30		0,0,-1			40,50,225 
+#  Rouge fond ver la camera
+pl 	0,0,-30		0,0,-1			255,0,0 
+# #  Rouge fond pas ver la camera
+# pl 	0,0,-30		0,0,1			255,0,0 
+
+# Blue droite vers la camera
+pl	8,0,0		1,0,0			0,0,255
+
+
+# pl 	0,0,-30		0,0,-1			255,0,0 
+
 # pl 	0,0,-30		0,0,-1			0,0,0 
 # pl 	0,-1,0		0,1,0			0,255,0 
 # pl	6,0,0			-1,0,0		255,32,43
