+++ scenes/cyk_face.rt
@@ -1,0 +1,12 @@
+A 0.2 255,255,255
+
+C 0,0,40 0,0,-1 90
+L 4,30,60 0.5
+
+sp 0,0,0 10  150,20,0
+
+cy 0,0,0 0,0,-1 5 50 0,20,150
+
+pl -5,0,0 1,0,0 100,100,100
+pl 0,-20,0 0,1,0 100,100,100
+pl 5,-20,0 1,0,0 100,100,100
