+++ scenes/pl_h.rt
@@ -1,11 +1,15 @@
+#	Ambient light
+#	intensity	color
 A	0.2							255,255,255
 
-C	0,400,-400		0,-1,-1	100
+#	Camera
+#	position	orientation		FOV
+C	0,400,-400		0,-1,+1	100
 L	-2,-2,-5				0.7
 
 sp	0,0,-5				1       	0,0,255
 sp	3,0,0				1			255,0,0
 sp	0,100.5,-1			200			0,255,0
 
-pl 10,10,10 1,0,0 255,0,225
+# pl 10,10,10 1,0,0 255,0,225
