+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: mpons <mpons@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/25 10:39:17 by slott             #+#    #+#             */
-/*   Updated: 2022/09/23 17:55:26 by mpons            ###   ########.fr       */
+/*   Updated: 2022/09/30 20:25:12 by mpons            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
